fix admin

+++ Model/ContentManager.php
@@ -120,7 +120,9 @@ class ContentManager
         $id = $_SESSION['user_id'];
         $plates['id'] = $data['idEditing'];
         $plateWithId = $this->getOnePlates($plates['id']);
-        unlink($plateWithId['image']);
+        if(!empty($img['fileEditing']['name'])){
+            unlink($plateWithId['image']);
+        }
         $plates['nameEditing'] = $data['plateNameEditing'];
         $plates['description'] = $data['descriptionEditing'];
         $plates['ingredients'] = $data['ingredientsEditing'];
@@ -144,12 +146,12 @@ class ContentManager
         if (empty($data['plateNameEditing'])  OR empty($data['descriptionEditing']) OR empty($data['allergenesEditing']) OR empty($data['ingredientsEditing']) OR empty($data['tricksEditing']) OR empty($data['priceEditing']) OR empty($data['categoryEditing'])) {
             return "Des champs obligatoire ne sont pas remplis";
         }
-    if(isset($img['fileEditing']['name']))){
-
-    }
+    if(!empty($img['fileEditing']['name'])){
         if (!in_array($extFile, $extension)) {
             return "Seul les images sont autoriséses";
         }
+    }
+
 
         return true;
     }
